Adjust spawn rates for the late dungeon
For almost all spawn rates, I have applied the following algorithm to
the start and end location for dungeon monster spawns:
   D:10 and earlier: no change
   D:11 --> D:10
   D:12 --> D:11
   D:13 --> D:11
   D:14 --> D:12
   D:15 --> D:12
   D:16 --> D:13
   D:17 --> D:13
   D:18 --> D:14
   D:19 --> D:15
   All later floors: subtract 4.

This may make out of depth spawns in the early dungeon slightly harsher,
but not affect general monster spawns up to D:10. Some playtesting will
be required to establish if this is the case.