Replace Ring of Flames with Ring of Steam (bhauth)
Ring of steam is a level 6 charms/fire/air spell that creates
steam clouds around the player, gives rF+ and rC- and reduces
player spell power (-1 enhancer level).

This spell is based on the Steam Barrier spell appearing in
bcrawl, but loses the AC bonus and gains air as an additional
school (the rationale being that blocking LoS with steam clouds
is a very powerful effect, so the spell needed a nerf).