Remove chunk eating for most species
Ghouls retain the ability because that's their thing. With spell hunger
removed food conducts become a vestigal part of crawl that can be used
to tweak the food clock for different species. Spriggans, with their
move speed, seem to be the only ones that require this adjustment,
accounted for by MUT_HERBIVORIOUS' nutrition penalty on rations.

Trolls lose their flavorful fast metabolism gourmand combo, which is a
shame and perhaps could be restored, but in practice it didn't confine
gameplay much for non-Gozag trolls. They retain fast metabolism 1 to
account for their regeneration.

Hydra form now allows devouring up to Engorged for all species that can
transmute.