Remove melee hunger
This is paired with the removal of spell hunger and chunk eating.
Abilities still have hunger, which makes their strategic costs more
relevant. If this is too relevant, they can lose hunger costs as well,
and the prevalence of rations adjusted to keep an ok food clock.

There have been reports that D:1-7 are a bit tight hunger wise even for
non-degenerate play; if melee hunger removal doesn't fully address this
an additional starting ration can be handed out.