Nerf Animate Skeleton and Animate Dead
Animate Skeleton (and to a slightly lesser extent Animate Dead) are
absurdly powerful: low level spells that grant depth-scaling long lived
allies with a summon cap determined only by the quantity of corpses
found. This power was granted independent of spellpower, so minimal
training was required to get full advantage of the spells. Miscasts put
a bit of a brake on AD, but because AS is level 1 this wasn't a serious
obstacle.

This commit addresses both the long-lived and power scaling issues.

Both spells now create minions with a summon duration of 4 (like
simulacrum) instead of 6, allowing them tactically and some amounts of
momentum but not the complete snowball of yore. Because of this duration
reduction, AS is freed from its location constraints, and raises a
single skeleton in LOS to make it more usable mid-fight.

Both spells now care about spell power and create their minions with hd
scaled by power, specifically minion_hd = mons_hd * (100 + pow) / 300. Animate
skeleton is capped at power 50, meaning an hd cap of 1/2 monster hd at
max power. In practice this means some early skeletons are HD 1 which might be
too severe, a lighter scaling could be applied, but AS is level 1 and
the skeletons retain the attack of their base monster. Animate Dead has
a power cap of 200, so with a training investment can reach a place
similar to its old power level; if this is too strong, a power cap of
150 can be applied.