Make plants wither away
Plants can create fun tactical situations. When they form
single-tile spaces in an otherwise-wider corridor, the player can
exploit that for a short-term advantage. That's fun. What's not fun
is getting a javelin stuck underneath a plant and feeling obliged to
slowly batter it to death with the cheapest means at your disposal.

So, refocus plants (and withered plants, demonic plants, fungi... but
not bushes) on this goal. When hit, they now get ENCH_SLOWLY_DYING,
killing them in 3-5 turns. Much cleaner