Isolate monster rng
There's lots of reasons to do this already, but after the recent poly
changes, the resulting rng state after every monster is dependent on the
global set of what monsters are valid, meaning that adding or removing
monsters would be highly likely to change every seed. So, isolate each
monster roll entirely so that stuff like this does not affect level
seeds.

Needless to say, this commit itself breaks seed stability for every
seed.