Reimagine Eringya's Noxious Bog (hellmonk,ebering)
The combined damage and movement/combat debuff of the bog is
interesting, but the spell's current approach of applying it in the
caster's wake makes the spell ineffective and somewhat difficult to use.
The additional turns spent moving to set up the effect incurs another
cost for a spell that's already damage over time. Obviously we could
increase the strength of the effect, but an alternative was proposed by
hellmonk: have the spell affect many cells, but only those not adjacent
to solid features. Being next to a lot of walls is generally helpful for
the player, so this requirement adds a different kind of cost, which is
that the player has to be near comparatively open areas. This is
positional, giving the spell a natural power scaling based on how open
the player's position is, and doesn't require kiting-like movement.

This commit changes the spell to transform all areas into bog that are
within a range of 4 and not adjacent to solid features, also excluding
the player's position. The transformation happens instantly and reverts
when the duration expires, with recasting not allowed until then. The
requirement that the feature have a solid floor remains (e.g. no deep
water or lava, but shallow water is fine). Additionally, only cells
visible via not-translocation-blocking LOS can be transformed, and
likewise we require that any transformed cell have all its adjacent
cells be visible. The second requirement is to prevent an information
leak by either using the targeter or casting the spell, either of which
would be a tedious exploit.

The duration is unchanged from before, as it seems reasonable for a
damage-over-time spell like this, but both this and the range of effect
of four will be reevaluated after more play-testing. If we want to
incorporate spell-power more, we one possible change is to increase the
range with spell-power, starting from a lower range at low power. We
have to strike a balance between the high level of the spell and its
need to be sufficiently strong and the fact that it can affect so many
more cells as we increase the range.

The spell has a static targeter showing potentially affected cells. I've
updated the casting noise from 2 to 6, which I assume was unintentional,
since we generally always have this noise value be the spell level. The
spell description is also updated. I've moved the spell from
spl-selfench.cc to spl-damage.cc, since this spell is really a damage
spell at its core, is similar to spells like Frozen Ramparts or Summon
Forest that not in spl-selfench.cc.