Un-de-enhance Scarf of Shadows spells
Per the original commit (bf280b6766):

    This gives 1 spell de-enhancer, since reduced LOS is very strong
    and allows more spells to be in range at the edge of LOS. If that
    makes this item too weak it can be removed.

The scarf of shadows effect is as strong or stronger for non-casting
characters as for casters, which made it odd that only the latter were
penalized for using it. Overall, after the scarf was changed to only
reduce LOS by 1 instead of 2, the effect was usually not strong enough
to justify both the loss of several AC (vs an enchanted cloak) *and* a
spell de-enhancer. Removing the latter should help make this a more
viable equipment choice.