Reimagine kobolds
Kobolds are a relatively uninteresting species. Giving kobolds nightstalker made them unique in mainline, but does not fit with traditional kobold lore. Instead, this commit gives kobolds the following:

* New (fake) mutation - redirection: Kobolds have a chance to redirect all dodged melee attacks in a given turn.

* 3 levels of the fangs mutation

* Adjusted aptitudes, including a penalty to all ranged skills and most magical skills, a larger gap between dodging and armour, and an increased long blades aptitude.

The goal of this reimagining is to create the experience of an EV tank provided by spriggans, but on a species with normal movement speed that excels at melee.