Make invisibility a level 5 charms/hexes spell
Invisibility has been a pure hexes spell largely for historical reasons - when enchantments split into charms and hexes, charms got haste and hexes got invisibility - but as a self-buff spell, it also fits thematically in the charms school. Furthermore, being in a spell school with limited crossover with other schools makes branching out to get invisibility rare for non-enchanters. Therefore, to better fit its theme and encourage non-enchanters to cast it, invisibility gains charms as a second school and has its level reduced to 5. This reduction of level still leads to a decrease in xp costs, despite making it dual-school (some rough calculations assuming 0 aptitudes estimate the cost at 1000 xp, equivalent to the cost of taking a skill from 0 to 6), leading to a small buff to a spell that is already quite strong; however, the preceeding upsides outweigh this downside."