Change Ozocubu's Armour's downside
Ozocubu's armour was made to end after moving to disincentivize always having the spell active, but this was a tedious mechanic - the player either had to be careful to not accidentally move when using tab, or deal with frequent prompts. This commit makes Ozocubu's armour instead start melting after a set period of time - that is, the AC bonus begins gradually decreasing to zero (at which point the spell ends). The player cannot recast the spell during this gradual weakening of the effect, thereby disincentivizing casting the spell out of combat, without the aforementioned issues.