Shorten Hell branches to 5 floors
Excessively long Hell branches contribute to making the extended
end game tedious. This commit therefore shortens the length of
Hell branches from 7 to 5. Monster spawn tables remain unchanged.
The frequency of Hell effects is increased by 1/3 on average
to compensate for the decreased depth, with the goal of
maintatining or slightly increasing the difficulty of the branch.