Shorten the dungeon to 12 floors (NormalPerson7)
The late dungeon is one of the easiest and least interesting parts of
the game, and by removing it we force players to enter lair branches at
the later part of XL 14, and also enter lair at the boundary of XL 11,
because it will now be risky to do D:10 and completely unsafe to do
D:11.

This commit removes D:13-15 in their entirety and changes the locations
in which branch entrances are found:
  - Lair now spawns on D:7-10 instead of D:8-11
  - Orc now spawns on D:8-11 instead of D:9-12
  - Vaults now spawns on D:11-12 instead of D:13-14
  - Depths now spawns on D:12, instead of D:15.

A subsequent commit will change the spawn chances for late dungeon
enemies to give D:12 a similar difficulty to old D:15, D:11 a similar
difficulty to old D:13, and D:10 a similar difficulty to old D:11. I
intend to try to keep the spawn chances on D:1-9 as close to previously
as possible.