Shorten Lair to 4 floors
Lair gives the player an important opportunity to gain equipment
and experience in a branch that is less variable than the Dungeon.
However, due to this lower variability, the gameplay in this
branch can become repetitive. Therefore we shorten Lair to 4 floors.

Inspired by NormalPerson7's approach to shortening the Dungeon, I
have applied the following algorithm for determining vault and
monster placements:
* L:1-2: no change
* L:3-5: subtract 1 floor
* L:6 and later: subtract 2 floors

This will make out-of-depths spawns on early Lair floors slightly
harsher, but this shouldn't be a major issue because Lair uses
mostly the same monsters on all floors, primarily just changing
their spawn weights.