Make the Forest ending more open
The map rot combined with the maziness of the modified Forest
layout made finding the end vault too tedious sometimes. Finding
the stairs however was not particularly problematic. To get the
best of both worlds, this commit uses the original unmodified
Forest layout for the brand ending, and the new modified version
for the higher level floors.