New Forest Monster and Effect: Yewborn (MalcolmRose)
Yewborn are slow tree monsters who can crush their foes with a
reaching branch attack and a constricting root attack.

Yewborn do not spawn as normal in the Forest; instead, every
100-200 turns up to 3 trees in LOS turn into yewborn (collecting
the mossy rune prevents this from happening). The goals of this
effect are to put a time pressure on the player and to make the
Forest feel more enchanted.