Remove Deep Dwarf
Deep dwarf's dwarviness and no regeneration are now shared by
sludge dwarf and vampire, respectively, making deep dwarf somewhat
redundant. On top of this, their handling of the no regen conduct
is poorly implemented for this species, pigeonholing the player
into selecting from a small list of gods, and leading to an
extremely overpowered species when the player chooses from this
list (most notably Makhleb). We therefore chose to remove this
species.