Remove almost all Forest loot outside of the end vaults
The map rot and yewborn spawns encourage players to dive to the
Forest end vault. Therefore, placing items in Forest outside of the
end vault encourages the player to backtrack after collecting the
rune, which is an anti-climactic end to the player's time in
Forest.

This commit adds the no_items branch flag to Forest and
removes most items appearing in non-end vaults. Thematically
appropriate (but not valuable) items are kept in some vaults,
and gammafunk_forest_yreds_foothold keeps its high-quality loot
(for now at least).