Reworked some costs for trove
Trove costs were too high for some items. This commit is a first
balancing pass to reduce the range of quantity of items requested
for certain items.

Potion of mutations have been added as a new item that troves can
request (but not for mummies, ghouls, vampires)

Scroll of fog and fear go from 4-8 to 4-7.
Potion of curing goes from 14-20 to 12-17.
Potion of heal wounds goes from 10-14 to 8-11.
Scroll of blinking goes from 4-8 to 3-5.
Scroll of teleport goes from 12-18 to 11-15.
Potion of haste goes from 4-6 to 3-4.
Potion of mutations are new and are 4-7.