Potion redux (#3)
Rework the potion of brilliance and potion of agility

Potion of brilliance now acts as a wizardry + spell enhancer +
mp cost reduction. 0.25 changed potion of brilliance and removed
the int bonus, spellpower bonus, and wizadry effect. In return,
it removed spell hunger from the spell and added a spell enhancer.

This commit reverts the wizardry bonus, but keeps the other changes
made in 0.25. With the removal of spell hunger, the effect is now
useless on the potion. 0.26 added halving mp costs as an effect
although it did not explicitly state spell hunger removal as the
rationale.

Potions of agility have returned and been reworked.
These now give +5 EV and temporary rsml instead of the +dex
they used to have. Hopefully this leads to more interesting
tactical uses of the potion.