Replace dazzling flash with battlesphere in the book of Conjurations
Now that dazzling flash does no damage and loses its conjurations
component, it makes little sense in the Conjurer's starter book.

Although battlesphere tended to "overshadow" fulminant prism when
it was in the book of Conjurations previously, it is a fun spell
that fits best in the early game, and ignoring fulminant prism was
never a good idea (it is less user friendly but arguably a
stronger spell).