Reimplement and rework Lava Orcs
Lava orcs make their triumphant return! Their mutations depend
on their temperature (a function of tension), and have been
implemented as follows:
* "Min" temperature:  rF+, swim through lava, earth magic boost,
                      slow movement (1.2 base move delay),
                      stone skin (AC:+2+XL/5);
* "Cold" temperature: Earth magic boost fades;
* "Warm" temperature: rF++, normal movement speed,
                      stone skin melts;
* "Hot" temperature:  rF+++, rC-, fire magic boost;
* "Fire" temperature: Passive heat (like Ds's passive freeze, but
                      reduced damage that scales with XL)
* "Max" temperature:  Cannot read scrolls, burn surroundings
                      (adjacent monsters take unavoidable fire
                      damage - currently the same damage as
                      passive heat).
The most notable differences from the old implementation are the
following:
* Damage from passive heat and burn surroundings scales with XL
  (1d(4+XL/5)-1 fire damage);
* Stoneskin scales with XL (rather than earth magic skill);
* Burn surroundings damage is substantially reduced, especially
  at low XL's;
* Slow movement is implemented without fast movement at high
  temperatures;
* Lava magic boost is replaced by earth magic boost, but this
  boost is only active at the lowest temperature tier;
* Most aptitudes are decreased by 1.
Thank you to shummie for getting the old lava orc temperature bar
code to work with the new local tiles UI; to Eronarn for the
concept and first implementation of the species; to kilobyte and
others for working on the later implentations; and Pete Hurst for
the making the new graphical features work on webtiles.